+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added self-hosted Urbanist web font [#59](https://github.com/interalia-studio/fullscreen/pull/59)
 - Added more tools to the new toolbar [#61](https://github.com/interalia-studio/fullscreen/pull/61)
 - Added Typescript config [#64](https://github.com/interalia-studio/fullscreen/pull/64)
+- Onboarding dialogue when joining a board on the web [#66](https://github.com/interalia-studio/fullscreen/pull/66)
 
 ### Changed